LayoutTests/media/track/track-cues-missed-expected.txt

Tests that events are triggered for missed (skipped) cues during normal playback.

EVENT(canplaythrough)
EXPECTED (testTrack.track.cues.length == '6') OK
RUN(video.play())
EVENT(enter)
EXPECTED (testTrack.track.cues.getCueById(cueCount) == '[object TextTrackCue]') OK
EXPECTED (currentCue.id == '3') OK
EVENT(exit)
EXPECTED (testTrack.track.cues.getCueById(cueCount) == '[object TextTrackCue]') OK
EXPECTED (currentCue.id == '3') OK
EVENT(enter)
EXPECTED (testTrack.track.cues.getCueById(cueCount) == '[object TextTrackCue]') OK
EXPECTED (currentCue.id == '4') OK
EVENT(exit)
EXPECTED (testTrack.track.cues.getCueById(cueCount) == '[object TextTrackCue]') OK
EXPECTED (currentCue.id == '4') OK
EVENT(enter)
EXPECTED (testTrack.track.cues.getCueById(cueCount) == '[object TextTrackCue]') OK
EXPECTED (currentCue.id == '5') OK
EVENT(exit)
EXPECTED (testTrack.track.cues.getCueById(cueCount) == '[object TextTrackCue]') OK
EXPECTED (currentCue.id == '5') OK
EVENT(enter)
EXPECTED (testTrack.track.cues.getCueById(cueCount) == '[object TextTrackCue]') OK
EXPECTED (currentCue.id == '6') OK
EVENT(exit)
EXPECTED (testTrack.track.cues.getCueById(cueCount) == '[object TextTrackCue]') OK
EXPECTED (currentCue.id == '6') OK
EVENT(ended)
END OF TEST
